+++ app/(afterLogin)/messages/[userId]/_hooks/useMessageForm.ts
@@ -1,44 +1,24 @@
-import toast from 'react-hot-toast';
 import { KeyboardEvent } from 'react';
-import { useMutation } from '@tanstack/react-query';
 import { SubmitHandler, useForm } from 'react-hook-form';
 
 import { IMessage } from '@/app/_types/message';
-import useOnAuth from '@/app/_hooks/useOnAuth';
-import { sendMessage } from '@/app/(afterLogin)/messages/[userId]/_services/sendMessage';
+import useSendMessage from '@/app/(afterLogin)/messages/[userId]/_hooks/useSendMessage';
 
 interface IUseMessageForm {
   receiverId: string;
 }
 
 const useMessageForm = ({ receiverId }: IUseMessageForm) => {
-  const { user } = useOnAuth();
   const {
     register,
     handleSubmit,
     reset,
     formState: { isSubmitting },
   } = useForm<IMessage>({ mode: 'onSubmit' });
+  const { sendToMessage } = useSendMessage({ receiverId, reset });
 
-  const sendToMessage = useMutation({
-    mutationFn: (data: IMessage) =>
-      sendMessage({
-        senderId: user?.displayName as string,
-        receiverId,
-        content: data.content,
-      }),
-    onSuccess: () => {
-      reset();
-      toast.success('메시지 보내기 성공');
-    },
-    onError: error => {
-      console.error('Error sending message:', error);
-      toast.error('메시지 보내기 실패');
-    },
-  });
-
-  const onSubmit: SubmitHandler<IMessage> = data => {
-    sendToMessage.mutate(data);
+  const onSubmit: SubmitHandler<IMessage> = async data => {
+    await sendToMessage.mutateAsync(data);
   };
 
   const handleKeyDown = (e: KeyboardEvent<HTMLTextAreaElement>) => {
